Shaders: Outlines and colors more visible
Changed the virtual object shadergraph so the colors are more visible
by multiplying the alpha of output of the emission to better see colors.

Increased the power of the fresnel effect so it doesn't ruin watching
colors from the side

Added an experimental check to see if the normal is up (y+) to give it a
solid color this needs some more improvements but helps with seeing
where the top of the cube is.